Extract raycaster
